rename: issue 파일을 context 폴더로 이동, 다른 파일에 경로 적용
